# Add additional benchmark for is_reachable #8155