@rossbar 2025-07-29
@Yasserelhaddar thank you for your contributions. I would just like to leave a personal note here: while there's no official policy re: the use of AI-generated pull requests, I will say as a reviewer these types of pull requests are not inspiring and often take more effort to review than they are worth. If you're going to use code gen tooling, I'd at least encourage you to see if you can't identify the core of the idea from the generated PR and simplify both the PR message and the code itself into a minimal, clear and concise proposal for a change. IMO this is a very valuable skill worth cultivating.

Anyways - I don't intend to discourage you from contributing to open source, I'd just like to share my perspective as a reviewer of contributions. Note that everything here is my personal perspective and not that of the project or any other organization/entity!