@Yasserelhaddar 2025-07-29
> IMO this is far too much code for a simple benchmark. This can be as simple as finding a tournament graph, even one that's randomly generated, then determining a pair of nodes where one is reachable from the other and another where this isn't the case, and add those as benchmarks.

On this one, I think tournament graphs have varying connectivity, some are strongly connected (no unreachable pairs), others have multiple SCCs with truly unreachable pairs. This creates inconsistent benchmark results.

We could switch to path graphs for guaranteed deterministic reachable/unreachable pairs, or test "direct edge" vs "indirect path" performance instead. Another option is using different graph types that guarantee unreachable pairs. Or we could just accept that some benchmarks may not run when no unreachable pairs exist.

What do you think ?