@dschult 2025-07-30
I think the suggestion was to find a single (could be randomly generated with a "seed" so it is the same every time the test is run) tournament graph long with a pair of nodes that are reachable and another pair of nodes that aren't reachable.

That's the hard work. Then the code just uses that graph and those two pairs of nodes to make sure the computed answer is correct.