@Yasserelhaddar 2025-08-05
Hi @dschult,

Thanks for the suggestion! I tried that approach, but I ran into a mathematical issue that I actually highlighted in my previous comment. When I searched for seeds that produce tournament graphs with unreachable pairs, I found that:

- **10 nodes**: Found `seed=1` works (has unreachable pairs) 
- **100 nodes**: No seed works - tournament graphs are strongly connected 
- **1000 nodes**: No seed works - even more strongly connected

By the time you reach 100+ nodes, almost all tournament graphs are strongly connected, making it impossible to find unreachable pairs for benchmarking.

**The fix I'm thinking of:**
Instead of tournament graphs, I'm using **DAGs (Directed Acyclic Graphs)** with a layered structure. This guarantees unreachable pairs at any size:

```python
def _make_benchmark_graphs_with_unreachable(seed):
    """Create graphs with guaranteed unreachable pairs using DAGs."""
    # Create 3 layers: nodes in layer 1 can't reach nodes in layer 3
    # This gives us guaranteed unreachable pairs for any graph size
```

As I mentioned in my previous comment, I tested this extensively and found that tournament graphs just get more connected as they get bigger. The DAG approach solves this mathematical limitation while still providing good benchmark coverage.

Would this DAG approach work for you? It addresses the connectivity issue I raised earlier.

Note that currently we try to construct and test 10, 100, 1000 Nodes Tournament Graphs:
```python
def _make_tournament_benchmark_graphs(seed):
    number_of_nodes = [10, 100, 1000]
    graphs = {}
    for nodes in number_of_nodes:
        G = nx.tournament.random_tournament(nodes, seed=seed)
        graphs[f"Tournament ({nodes}, seed={seed})"] = G
    return graphs
```

Thanks!