@dschult 2025-08-07
This approach seems overly complex to me -- but you haven't provided any code for the DAG method.
The current code in this PR is also overly complex. We don't need to compute all the pairs that are in the same scc (or in different sccs). We only need one pair for each case.